Fixed output so date comes first
Excel doesn't plot well when the x axis value is in the right hand column. This commit moves date to the first column and removes the newline that ctime inserts. It also rearranges the output so that it can more easily be piped to csv for plotting later by writing the header first then outputting the data in a loop.